Rename db module
